Module Context.Client_state

type t

val find : Tezos_client_019_PtParisB.Protocol_client_context.full -> Tezos_protocol_019_PtParisB.Protocol.Contract_hash.t -> t -> Contract_state.t Tezos_base.TzPervasives.tzresult Lwt.t

val register : Tezos_client_019_PtParisB.Protocol_client_context.full -> force:bool -> default_memo_size:int option -> Tezos_protocol_019_PtParisB.Protocol.Contract_hash.t -> Tezos_sapling.Core.Client.Viewing_key.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val sync_and_scan : Tezos_client_019_PtParisB.Protocol_client_context.full -> Tezos_protocol_019_PtParisB.Protocol.Contract_hash.t -> Contract_state.t Tezos_base.TzPervasives.tzresult Lwt.t

Synchronise our local state with the blockchain's. The state must be recent enough to craft correct transactions. The limit enforced by the protocol if 120 blocks. Also scans, ie. checks for incoming payments and add them to our balance. *