Module RPC.Scripts

module S : sig ... end

module type UNPARSING_MODE = sig ... end

module Traced_interpreter (Unparsing_mode : UNPARSING_MODE) : sig ... end

module Unparse_types : sig ... end

module Normalize_stack : sig ... end

val pp_instr_name : 'a 'b 'c 'd. Tezos_protocol_019_PtParisB.Environment.Format.formatter -> ('a, 'b, 'c, 'd) Tezos_protocol_019_PtParisB.Protocol.Script_typed_ir.kinstr -> unit

type Tezos_protocol_019_PtParisB.Environment.Error_monad.error +=

| Run_operation_does_not_support_consensus_operations

val run_operation_service : Tezos_protocol_019_PtParisB.Protocol.Services_registration.rpc_context -> < version : 'a.. > -> (Tezos_protocol_019_PtParisB.Protocol.Alpha_context.packed_operation * Tezos_protocol_environment_019_PtParisB.Chain_id.t) -> ('a * (Tezos_protocol_019_PtParisB.Protocol.Alpha_context.packed_protocol_data * Tezos_raw_protocol_019_PtParisB.Apply_results.packed_operation_metadata), Tezos_protocol_019_PtParisB.Environment.Error_monad.error Tezos_protocol_environment_019_PtParisB.Error_monad.trace) Stdlib.result Lwt.t

Validate and apply the operation but skip signature checks; do not support consensus operations.

Return the unchanged operation protocol data, and the operation receipt ie. metadata containing balance updates, consumed gas, application success or failure, etc.

val simulate_operation_service : Tezos_protocol_019_PtParisB.Protocol.Services_registration.rpc_context -> < version : 'a.. > -> (int32 option * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.packed_operation * Tezos_protocol_environment_019_PtParisB.Chain_id.t * int) -> ('a * (Tezos_protocol_019_PtParisB.Protocol.Alpha_context.packed_protocol_data * Tezos_raw_protocol_019_PtParisB.Apply_results.packed_operation_metadata), Tezos_protocol_019_PtParisB.Environment.Error_monad.error Tezos_protocol_environment_019_PtParisB.Error_monad.trace) Stdlib.result Lwt.t

val default_from_context : 'a -> ('a -> ('b, 'c) Stdlib.result Lwt.t) -> 'b option -> ('b, 'c) Stdlib.result Lwt.t

type run_code_config = {

balance : Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Tez.t;

self : Tezos_protocol_019_PtParisB.Protocol.Contract_hash.t;

payer : Tezos_protocol_019_PtParisB.Environment.Signature.public_key_hash;

sender : Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Contract.t;

}

val default_balance : Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Tez.t

val register : unit -> unit

val run_code : unparsing_mode: Tezos_protocol_019_PtParisB.Protocol.Script_ir_unparser.unparsing_mode option -> gas: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.Arith.integral option -> entrypoint:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Entrypoint.t -> balance:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Tez.t option -> other_contracts:S.other_contract_description list option -> extra_big_maps:S.extra_big_map_description list option -> script:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> storage:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> input:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> amount:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Tez.t -> chain_id:Tezos_protocol_019_PtParisB.Environment.Chain_id.t -> sender:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Contract.t option -> payer: Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t option -> self:Tezos_raw_protocol_019_PtParisB.Contract_hash.t option -> now:Tezos_protocol_019_PtParisB.Protocol.Script_timestamp.t option -> level: Tezos_protocol_019_PtParisB.Protocol.Script_int.n Tezos_protocol_019_PtParisB.Protocol.Script_int.num option -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> (Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr * Tezos_protocol_019_PtParisB.Protocol.Apply_internal_results.packed_internal_operation list * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Lazy_storage.diffs option) Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val trace_code : unparsing_mode: Tezos_protocol_019_PtParisB.Protocol.Script_ir_unparser.unparsing_mode option -> gas: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.Arith.integral option -> entrypoint:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Entrypoint.t -> balance:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Tez.t option -> other_contracts:S.other_contract_description list option -> extra_big_maps:S.extra_big_map_description list option -> script:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> storage:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> input:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> amount:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Tez.t -> chain_id:Tezos_protocol_019_PtParisB.Environment.Chain_id.t -> sender:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Contract.t option -> payer: Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t option -> self:Tezos_raw_protocol_019_PtParisB.Contract_hash.t option -> now:Tezos_protocol_019_PtParisB.Protocol.Script_timestamp.t option -> level: Tezos_protocol_019_PtParisB.Protocol.Script_int.n Tezos_protocol_019_PtParisB.Protocol.Script_int.num option -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> (Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr * Tezos_protocol_019_PtParisB.Protocol.Apply_internal_results.packed_internal_operation list * Tezos_protocol_019_PtParisB.Protocol.Script_typed_ir.execution_trace * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Lazy_storage.diffs option) Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val run_tzip4_view : gas: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.Arith.integral option -> other_contracts:S.other_contract_description list option -> extra_big_maps:S.extra_big_map_description list option -> contract:Tezos_raw_protocol_019_PtParisB.Contract_hash.t -> entrypoint:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Entrypoint.t -> input:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> chain_id:Tezos_protocol_019_PtParisB.Environment.Chain_id.t -> now:Tezos_protocol_019_PtParisB.Protocol.Script_timestamp.t option -> level: Tezos_protocol_019_PtParisB.Protocol.Script_int.n Tezos_protocol_019_PtParisB.Protocol.Script_int.num option -> sender:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Contract.t option -> payer: Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t option -> unparsing_mode: Tezos_protocol_019_PtParisB.Protocol.Script_ir_unparser.unparsing_mode -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val run_script_view : gas: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.Arith.integral option -> other_contracts:S.other_contract_description list option -> extra_big_maps:S.extra_big_map_description list option -> contract:Tezos_raw_protocol_019_PtParisB.Contract_hash.t -> view:string -> input:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> unlimited_gas:bool -> chain_id:Tezos_protocol_019_PtParisB.Environment.Chain_id.t -> now:Tezos_protocol_019_PtParisB.Protocol.Script_timestamp.t option -> level: Tezos_protocol_019_PtParisB.Protocol.Script_int.n Tezos_protocol_019_PtParisB.Protocol.Script_int.num option -> sender:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Contract.t option -> payer: Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t option -> unparsing_mode: Tezos_protocol_019_PtParisB.Protocol.Script_ir_unparser.unparsing_mode -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

run_script_view is an helper function to call the corresponding RPC.

val run_instr : gas: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.Arith.integral option -> legacy:bool -> input: (Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr) list -> code:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> chain_id:Tezos_protocol_019_PtParisB.Environment.Chain_id.t -> now:Tezos_protocol_019_PtParisB.Protocol.Script_timestamp.t option -> level: Tezos_protocol_019_PtParisB.Protocol.Script_int.n Tezos_protocol_019_PtParisB.Protocol.Script_int.num option -> unparsing_mode: Tezos_protocol_019_PtParisB.Protocol.Script_ir_unparser.unparsing_mode option -> source: Tezos_protocol_environment_019_PtParisB.Signature.Public_key_hash.t option -> sender:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Contract.t option -> self:Tezos_raw_protocol_019_PtParisB.Contract_hash.t option -> parameter: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr option -> amount:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Tez.t -> balance:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Tez.t option -> other_contracts:S.other_contract_description list option -> extra_big_maps:S.extra_big_map_description list option -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> ((Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr) list * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.t) Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val typecheck_code : gas: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.Arith.integral option -> legacy:bool -> script:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> show_types:bool -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> ((Tezos_raw_protocol_019_PtParisB.Alpha_context.Script.location * (Tezos_raw_protocol_019_PtParisB.Alpha_context.Script.expr list * Tezos_raw_protocol_019_PtParisB.Alpha_context.Script.expr list)) list * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.t) Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val script_size : gas: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.Arith.integral option -> legacy:bool -> script:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> storage:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> int Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val typecheck_data : gas: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.Arith.integral option -> legacy:bool -> data:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> ty:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.t Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val pack_data : gas: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.Arith.integral option -> data:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> ty:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> (Tezos_protocol_environment_019_PtParisB.Bytes.t * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Gas.t) Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val normalize_data : legacy:bool -> other_contracts:S.other_contract_description list option -> extra_big_maps:S.extra_big_map_description list option -> data:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> ty:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> unparsing_mode: Tezos_protocol_019_PtParisB.Protocol.Script_ir_unparser.unparsing_mode -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val normalize_stack : legacy:bool -> other_contracts:S.other_contract_description list option -> extra_big_maps:S.extra_big_map_description list option -> stack: (Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr) list -> unparsing_mode: Tezos_protocol_019_PtParisB.Protocol.Script_ir_unparser.unparsing_mode -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> (Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr) list Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val normalize_script : script:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> unparsing_mode: Tezos_protocol_019_PtParisB.Protocol.Script_ir_unparser.unparsing_mode -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val normalize_type : ty:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val run_operation : op:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Operation.packed -> chain_id:Tezos_protocol_019_PtParisB.Environment.Chain_id.t -> ?version:version -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> (Tezos_raw_protocol_019_PtParisB.Alpha_context.Operation.packed_protocol_data * Tezos_protocol_019_PtParisB.Protocol.Apply_results.packed_operation_metadata, Tezos_base.TzPervasives.Error_monad.tztrace) Stdlib.result Lwt.t

val simulate_operation : op:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Operation.packed -> chain_id:Tezos_protocol_019_PtParisB.Environment.Chain_id.t -> latency:int -> ?version:version -> ?successor_level:bool -> ?blocks_before_activation:int32 -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> (Tezos_raw_protocol_019_PtParisB.Alpha_context.Operation.packed_protocol_data * Tezos_protocol_019_PtParisB.Protocol.Apply_results.packed_operation_metadata, Tezos_base.TzPervasives.Error_monad.tztrace) Stdlib.result Lwt.t

val entrypoint_type : script:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> entrypoint:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Entrypoint.t -> 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val list_entrypoints : 'a Tezos_protocol_019_PtParisB.Environment.RPC_context.simple -> 'a -> script:Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr -> (Tezos_protocol_019_PtParisB.Protocol.Michelson_v1_primitives.prim list list * (string * Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Script.expr) list) Tezos_protocol_environment_019_PtParisB.Error_monad.shell_tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t