Module Plugin_registerer.Http_cache_headers

include module type of struct include Tezos_protocol_plugin_019_PtParisB.Plugin.Http_cache_headers end

val get_round_end_time : get_context:(unit -> Tezos_protocol_environment.Context.t Lwt.t) -> Tezos_base.TzPervasives.Block_header.shell_header -> Tezos_base.TzPervasives.Time.System.t option Lwt.t

get_round_end_time get_context header gets the time at which the current round ends which is the time at which the next round starts. Useful to get an estimate of when the next block should arrive. get_context is called the first time this function is called to get the current constants.

val hash : Tezos_base.TzPervasives.Protocol_hash.t