Module Game_helpers.Wasm

val new_dissection : default_number_of_sections:int -> start_chunk: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Sc_rollup.Game.dissection_chunk -> our_stop_chunk: Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Sc_rollup.Game.dissection_chunk -> Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Sc_rollup.Tick.t list

new_dissection ~default_number_of_sections ~start_chunk ~our_stop_chunk computes a dissection that satisfies the dissection predicate of the WASM PVM, that is all the ticks in the dissection are aligned with the size of a snapshot.

If start_chunk is not a multiple of the size of a snapshot or if the distance between start_chunk and stop_chunk is not greater than the size of a snapshot, then default_new_dissection is called, because it means the WASM PVM is stuck.