Module Pvm_in_memory.Wasm

Wasm Wasm PVM with an in memory context Tezos_context_memory.

include Tezos_protocol_019_PtParisB.Protocol.Alpha_context.Sc_rollup.PVM.S with type context = Tezos_context_memory.Context_binary.t and type state = Tezos_context_memory.Context_binary.tree and type proof = Tezos_context_memory.Context_binary.Proof.tree Tezos_context_memory.Context_binary.Proof.t

val parse_boot_sector : string -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.PVM.boot_sector option

val pp_boot_sector : Tezos_protocol_environment_019_PtParisB.Format.formatter -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.PVM.boot_sector -> unit

type state = Tezos_context_memory.Context_binary.tree

val pp : state -> (Tezos_protocol_environment_019_PtParisB.Format.formatter -> unit -> unit) Tezos_protocol_environment_019_PtParisB.Lwt.t

type context = Tezos_context_memory.Context_binary.t

type hash = Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.State_hash.t

type proof = Tezos_context_memory.Context_binary.Proof.tree Tezos_context_memory.Context_binary.Proof.t

val proof_encoding : proof Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val proof_start_state : proof -> hash

val proof_stop_state : proof -> hash

val state_hash : state -> hash Tezos_protocol_environment_019_PtParisB.Lwt.t

val initial_state : empty:state -> state Tezos_protocol_environment_019_PtParisB.Lwt.t

val install_boot_sector : state -> string -> state Tezos_protocol_environment_019_PtParisB.Lwt.t

val is_input_state : is_reveal_enabled: Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.is_reveal_enabled -> state -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.input_request Tezos_protocol_environment_019_PtParisB.Lwt.t

val set_input : Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.input -> state -> state Tezos_protocol_environment_019_PtParisB.Lwt.t

val eval : state -> state Tezos_protocol_environment_019_PtParisB.Lwt.t

val verify_proof : is_reveal_enabled: Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.is_reveal_enabled -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.input option -> proof -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.input_request Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val produce_proof : context -> is_reveal_enabled: Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.is_reveal_enabled -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.input option -> state -> proof Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

type output_proof

val output_proof_encoding : output_proof Tezos_protocol_environment_019_PtParisB.Data_encoding.t

val output_of_output_proof : output_proof -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.output

val state_of_output_proof : output_proof -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.State_hash.t

val verify_output_proof : output_proof -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.output Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult Tezos_protocol_environment_019_PtParisB.Lwt.t

val produce_output_proof : context -> state -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.output -> (output_proof, Tezos_protocol_environment_019_PtParisB.Error_monad.error) Tezos_protocol_environment_019_PtParisB.Pervasives.result Tezos_protocol_environment_019_PtParisB.Lwt.t

val check_dissection : default_number_of_sections:int -> start_chunk: Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.Dissection_chunk.t -> stop_chunk: Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.Dissection_chunk.t -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Sc_rollup.Dissection_chunk.t list -> unit Tezos_protocol_environment_019_PtParisB.Error_monad.tzresult

val get_current_level : state -> Tezos_raw_protocol_019_PtParisB.Alpha_context.Raw_level.t option Tezos_protocol_environment_019_PtParisB.Lwt.t

module Internal_for_tests : sig ... end

val make_empty_context : unit -> context

make_empty_context () create a new in memory context for a PVM.

val make_empty_state : unit -> state

make_empty_state () create a new state with an in memory context for a PVM.